+++ src/config.rs
@@ -4,7 +4,10 @@ use std::{
 };
 
 #[cfg(doc)]
-use crate::protocols::{self, Handshake, Reading, Writing};
+use crate::{
+    protocols::{self, Handshake, Reading, Writing},
+    Node,
+};
 
 /// The node's configuration. See the source of [`Config::default`] for the defaults.
 #[derive(Debug, Clone)]