+++ tests/disconnect.rs
@@ -30,8 +30,7 @@ async fn connector_side_disconnect() {
     let connectee = crate::test_node!("connectee0");
     connectee.enable_reading().await;
     connectee.enable_disconnect().await;
-
-    let connectee_addr = connectee.node().listening_addr().unwrap();
+    let connectee_addr = connectee.node().start_listening().await.unwrap();
 
     connector.node().connect(connectee_addr).await.unwrap();
 
@@ -65,8 +64,7 @@ async fn connectee_side_disconnect() {
     let connectee = crate::test_node!("connectee1");
     connectee.enable_writing().await;
     connectee.enable_disconnect().await;
-
-    let connectee_addr = connectee.node().listening_addr().unwrap();
+    let connectee_addr = connectee.node().start_listening().await.unwrap();
 
     connector.node().connect(connectee_addr).await.unwrap();