+++ tests/node.rs
@@ -19,12 +19,12 @@ use pea2pea::{connect_nodes, protocols::Handshake, Config, Node, Pea2Pea, Topolo
 
 #[tokio::test]
 async fn node_creation_any_port_works() {
-    let _node = Node::new(Default::default()).await.unwrap();
+    let _node = Node::new(Default::default());
 }
 
 #[tokio::test]
 async fn node_name_gets_auto_assigned() {
-    let node = Node::new(Default::default()).await.unwrap();
+    let node = Node::new(Default::default());
     // ensure that a node without a given name get assigned a numeric ID
     let _: usize = node.config().name.as_ref().unwrap().parse().unwrap();
 }
@@ -35,21 +35,22 @@ async fn node_given_name_remains_unchanged() {
         name: Some("test".into()),
         ..Default::default()
     };
-    let node = Node::new(config).await.unwrap();
+    let node = Node::new(config);
     // ensure that a node with a given name doesn't have it overwritten
     assert_eq!(node.config().name.as_ref().unwrap(), "test");
 }
 
 #[tokio::test]
 async fn node_use_provided_socket() {
-    let connector = Node::new(Default::default()).await.unwrap();
-    let connectee = Node::new(Default::default()).await.unwrap();
+    let connector = Node::new(Default::default());
+    let connectee = Node::new(Default::default());
+    let connectee_addr = connectee.start_listening().await.unwrap();
 
     let socket = TcpSocket::new_v4().unwrap();
     socket.bind("127.0.0.77:0".parse().unwrap()).unwrap();
 
     connector
-        .connect_using_socket(connectee.listening_addr().unwrap(), socket)
+        .connect_using_socket(connectee_addr, socket)
         .await
         .unwrap();
 
@@ -64,6 +65,23 @@ async fn node_use_provided_socket() {
     );
 }
 
+#[tokio::test]
+async fn node_delayed_listener() {
+    let connector = Node::new(Default::default());
+    let connectee = Node::new(Default::default());
+
+    assert!(connectee.listening_addr().is_err());
+
+    let connectee_addr = connectee.start_listening().await.unwrap();
+
+    connector.connect(connectee_addr).await.unwrap();
+
+    let connectee_clone = connectee.clone();
+    deadline!(Duration::from_secs(1), move || connectee_clone
+        .num_connected()
+        == 1);
+}
+
 #[should_panic]
 #[tokio::test]
 async fn node_creation_bad_params_panic() {
@@ -71,7 +89,8 @@ async fn node_creation_bad_params_panic() {
         allow_random_port: false,
         ..Default::default()
     };
-    let _node = Node::new(config).await.unwrap();
+    let node = Node::new(config);
+    node.start_listening().await.unwrap();
 }
 
 #[tokio::test]
@@ -81,7 +100,9 @@ async fn node_creation_used_port_fails() {
         allow_random_port: false,
         ..Default::default()
     };
-    assert!(Node::new(config).await.is_err());
+    let node = Node::new(config);
+
+    assert!(node.start_listening().await.is_err());
 }
 
 #[tokio::test]
@@ -145,8 +166,9 @@ async fn node_connecting() {
 
 #[tokio::test]
 async fn node_self_connection_fails() {
-    let node = Node::new(Default::default()).await.unwrap();
-    assert!(node.connect(node.listening_addr().unwrap()).await.is_err());
+    let node = Node::new(Default::default());
+    let own_addr = node.start_listening().await.unwrap();
+    assert!(node.connect(own_addr).await.is_err());
 }
 
 #[tokio::test]
@@ -170,13 +192,11 @@ async fn node_connector_limit_breach_fails() {
         max_connections: 0,
         ..Default::default()
     };
-    let connector = Node::new(config).await.unwrap();
-    let connectee = Node::new(Default::default()).await.unwrap();
+    let connector = Node::new(config);
+    let connectee = Node::new(Default::default());
+    let connectee_addr = connectee.start_listening().await.unwrap();
 
-    assert!(connector
-        .connect(connectee.listening_addr().unwrap())
-        .await
-        .is_err());
+    assert!(connector.connect(connectee_addr).await.is_err());
 }
 
 #[tokio::test]
@@ -185,14 +205,13 @@ async fn node_connectee_limit_breach_fails() {
         max_connections: 0,
         ..Default::default()
     };
-    let connectee = Node::new(config).await.unwrap();
-    let connector = Node::new(Default::default()).await.unwrap();
+    let connectee = Node::new(config);
+    let connectee_addr = connectee.start_listening().await.unwrap();
+
+    let connector = Node::new(Default::default());
 
     // a breached connection limit doesn't close the listener, so this works
-    connector
-        .connect(connectee.listening_addr().unwrap())
-        .await
-        .unwrap();
+    connector.connect(connectee_addr).await.unwrap();
 
     // the number of connections on connectee side needs to be checked instead
     deadline!(Duration::from_secs(1), move || connectee.num_connected()
@@ -203,16 +222,17 @@ async fn node_connectee_limit_breach_fails() {
 async fn node_overlapping_duplicate_connection_attempts_fail() {
     const NUM_ATTEMPTS: usize = 5;
 
-    let connector = Node::new(Default::default()).await.unwrap();
-    let connectee = Node::new(Default::default()).await.unwrap();
-    let addr = connectee.listening_addr().unwrap();
+    let connector = Node::new(Default::default());
+
+    let connectee = Node::new(Default::default());
+    let connectee_addr = connectee.start_listening().await.unwrap();
 
     let err_count = Arc::new(AtomicUsize::new(0));
     for _ in 0..NUM_ATTEMPTS {
         let connector_clone = connector.clone();
         let err_count_clone = err_count.clone();
         tokio::spawn(async move {
-            if connector_clone.connect(addr).await.is_err() {
+            if connector_clone.connect(connectee_addr).await.is_err() {
                 err_count_clone.fetch_add(1, Relaxed);
             }
         });
@@ -224,8 +244,8 @@ async fn node_overlapping_duplicate_connection_attempts_fail() {
 
 #[tokio::test]
 async fn node_shutdown_closes_the_listener() {
-    let node = Node::new(Default::default()).await.unwrap();
-    let addr = node.listening_addr().unwrap();
+    let node = Node::new(Default::default());
+    let addr = node.start_listening().await.unwrap();
 
     assert!(TcpListener::bind(addr).await.is_err());
     node.shut_down().await;
@@ -235,7 +255,8 @@ async fn node_shutdown_closes_the_listener() {
 
 #[tokio::test]
 async fn test_nodes_use_localhost() {
-    let node = Node::new(Default::default()).await.unwrap();
+    let node = Node::new(Default::default());
+    let addr = node.start_listening().await.unwrap();
 
-    assert_eq!(node.listening_addr().unwrap().ip(), Ipv4Addr::LOCALHOST);
+    assert_eq!(addr.ip(), Ipv4Addr::LOCALHOST);
 }